pyproject.toml
[project]
name = "spotify_api_client"
version = "0.1.0"
description = "Low-level spotify api client."
readme = "https://github.com/richardpordan/spotify-api-client/README.md"
requires-python = ">=3.13"

dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "isort>=8.0.1",
    "prek>=0.3.6",
    "ruff>=0.15.7",
]
test = [
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RET", # flake8-return
    "ARG", # unused arguments
    "PTH", # use pathlib
    "RUF", # Ruff-specific rules
    "PL",  # pylint rules
    "S",   # security
    "T",   # Type hints
    "D",   # Docstring style
    "COM", # Complexity checks
    "PYI", # stub file checks
]

ignore = ["COM812", "N802"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[lint.ruff.per-file-ignores]
"tests/*.py" = [
    "S101",  # allow asserts in tests
]
